requirements.txt
# Core dependencies for the chords project
pyfluidsynth>=1.3.0
mido>=1.2.0
python-rtmidi>=1.4.0

# Jupyter notebook support
jupyter>=1.0.0
notebook>=6.0.0

# Audio processing (optional but recommended)
numpy>=1.20.0

# Additional useful packages
ipykernel>=6.0.0

# Web framework
flask>=2.0.0
